move "validate_sequences" code from bids_validator.py to validation.py
